created the ability to show all users, edited the create_user function, and added plans to the readme

+++ users/crud.py
@@ -1,33 +1,54 @@
-from typing import Type
+from typing import Type, Sequence
 
 from fastapi import HTTPException
 from sqlalchemy.exc import SQLAlchemyError, NoResultFound
 from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy import select
+from sqlalchemy import select, Select, Result, func
 from datetime import datetime
 
 from core.models.models import User
-from .schemas import BaseUser, User
-from core.models import models
+from .schemas import BaseUser, CreateUser
 
 
 async def get_user(session: AsyncSession, user_id: int) -> Type[User] | None:
     try:
-        user = await session.get(models.User, user_id)
+        user = await session.get(User, user_id)
         if user is None:
             raise NoResultFound("User not found")
         return user
     except SQLAlchemyError as e:
         raise HTTPException(status_code=500, detail="Database error")
+    except Exception as e:
+        await session.rollback()
+        raise e
+    finally:
+        await session.close()
 
 
-async def create_user(session: AsyncSession, user_in: BaseUser) -> User:
-    user = models.User(
-        id=2,
-        registered_at=datetime.utcnow(),
-        **user_in.model_dump()
-    )
+async def get_users(session: AsyncSession) -> Sequence[User]:
     try:
+        users = await session.execute(select(User))
+        if users:
+            return users.scalars().all()
+        else:
+            raise HTTPException(status_code=404, detail="User not found")
+    except SQLAlchemyError as e:
+        raise HTTPException(status_code=500, detail="Database error")
+    except Exception as e:
+        await session.rollback()
+        raise e
+    finally:
+        await session.close()
+
+
+async def create_user(session: AsyncSession, user_in: BaseUser) -> CreateUser:
+    try:
+        count_users = await session.scalar(select(func.count()).select_from(User))
+        user = User(
+            id=count_users,
+            registered_at=datetime.utcnow(),
+            **user_in.model_dump()
+        )
         session.add(user)
         await session.commit()
     except SQLAlchemyError as e:
@@ -37,12 +58,12 @@ async def create_user(session: AsyncSession, user_in: BaseUser) -> User:
 
 async def delete_user(session: AsyncSession, user_id: int) -> None:
     try:
-        user = await session.get(models.User, user_id)
-
-        await session.delete(user)
-        await session.commit()
-    except NoResultFound:
-        raise HTTPException(status_code=404, detail="User not found")
+        user = await session.get(User, user_id)
+        if user:
+            await session.delete(user)
+            await session.commit()
+        else:
+            raise HTTPException(status_code=404, detail="User not found")
     except Exception as e:
         await session.rollback()
         raise e
@@ -57,16 +78,16 @@ async def update_user(
         partial: bool = False,
 ) -> Type[User]:
     try:
-        find_user = await session.get(models.User, user_id)
+        find_user = await session.get(User, user_id)
+        if user:
+            find_user.name = user.name
+            find_user.email = user.email
+            find_user.password = user.password
 
-        find_user.name = user.name
-        find_user.email = user.email
-        find_user.password = user.password
-
-        await session.commit()
-        return find_user
-    except NoResultFound:
-        raise HTTPException(status_code=404, detail="User not found")
+            await session.commit()
+            return find_user
+        else:
+            raise HTTPException(status_code=404, detail="User not found")
     except Exception as e:
         await session.rollback()
         raise e
